# contracts/ingestion-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Canal — Realtime ingestion API (Phase 1 MVP)
  version: 0.1.0
  description: |
    HTTP-first ingestion contract for Phase 1 MVP. Streaming delivery (SSE/WebSocket)
    is reserved for a follow-on iteration; `GET /v1/stream` is a documented stub.

    Product pipeline stage names and ordering (including “no indentation semantics”)
    are frozen in company issue [CAN-30](/CAN/issues/CAN-30) issue document
    `data-architecture` §2.3 (rev 3); see also RFC
    `docs/rfc-phase1-buffer-checkpoint-provider-matrix.md` §1.5.

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Health
  - name: Ingestion
  - name: Control
  - name: AdapterInstances

paths:
  /health:
    get:
      tags: [Health]
      summary: Liveness / readiness probe
      operationId: getHealth
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /v1/events:
    post:
      tags: [Ingestion]
      summary: Accept a batch of raw ingestion events (idempotent by event id)
      operationId: postIngestEvents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestBatchRequest"
      responses:
        "202":
          description: Events accepted for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestBatchResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/stream:
    get:
      tags: [Ingestion]
      summary: Realtime fan-out (stub — not implemented in MVP scaffold)
      operationId: getIngestStream
      description: |
        Placeholder for future SSE/WebSocket subscription. Implementations SHOULD return
        `501` until streaming is shipped; clients MUST NOT depend on this path in Phase 1.
      responses:
        "501":
          description: Streaming not yet available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/control/pipeline:
    get:
      tags: [Control]
      summary: Pipeline summary and adapter placeholders for operator UI
      operationId: getControlPipeline
      description: |
        Read model listing the board-approved sequential pipeline (CAN-30 / RFC §1.5) and
        placeholder adapter instances with **catalog tier** metadata (`tier-1` … `tier-3`
        charter taxonomy). Deterministic scaffold data until connector registry is wired.
      responses:
        "200":
          description: Current pipeline read model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineSummaryRead"

  /v1/control/canal/segments:
    get:
      tags: [Control]
      summary: Canal buffer segments aligned to the pipeline topology
      operationId: getControlCanalSegments
      description: |
        Read model of logical **buffer** segments (source / core / sink event buffers)
        for wizard and operator diagrams. Provider profile is Phase 1 local only.
      responses:
        "200":
          description: Canal segment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanalSegmentsRead"

  /v1/adapter-instances:
    get:
      tags: [AdapterInstances]
      summary: List persisted operator adapter instance bindings
      operationId: listAdapterInstances
      description: |
        Phase 1 in-memory records. `catalogTier` is always derived from the pipeline
        read model (`GET /v1/control/pipeline`) for the bound `catalogAdapterId` —
        clients cannot set tier independently (CAN-51).
      responses:
        "200":
          description: Adapter instances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdapterInstanceListRead"
    post:
      tags: [AdapterInstances]
      summary: Create an adapter instance bound to a catalog placeholder
      operationId: createAdapterInstance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdapterInstanceCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdapterInstanceRead"
        "400":
          description: Validation error or unknown catalog adapter id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/adapter-instances/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [AdapterInstances]
      summary: Get one adapter instance
      operationId: getAdapterInstance
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdapterInstanceRead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [AdapterInstances]
      summary: Update operator label and/or rebind to another catalog placeholder
      operationId: patchAdapterInstance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdapterInstancePatchRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdapterInstanceRead"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [AdapterInstances]
      summary: Delete an adapter instance
      operationId: deleteAdapterInstance
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthResponse:
      type: object
      required: [status, service, version]
      properties:
        status:
          type: string
          enum: [ok]
        service:
          type: string
          example: ingestion-api
        version:
          type: string
          example: "0.1.0"

    IngestBatchRequest:
      type: object
      required: [source, events]
      properties:
        source:
          type: string
          description: Logical producer (connector id, app name, etc.)
          minLength: 1
          maxLength: 256
        events:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: "#/components/schemas/IngestEvent"

    IngestEvent:
      type: object
      required: [id, type, occurredAt]
      properties:
        id:
          type: string
          description: Idempotency key for deduplication
          minLength: 1
          maxLength: 128
        type:
          type: string
          description: Event type / schema discriminator
          minLength: 1
          maxLength: 256
        occurredAt:
          type: string
          format: date-time
        payload:
          description: Arbitrary JSON payload
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
            - type: string
            - type: number
            - type: boolean
            - type: "null"

    IngestBatchResponse:
      type: object
      required: [accepted, duplicateIds]
      properties:
        accepted:
          type: integer
          minimum: 0
        duplicateIds:
          type: array
          items:
            type: string

    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string

    CatalogTier:
      type: string
      description: >
        Charter / implementation taxonomy for ingestion path (wizard lanes). Distinct from
        commercial `ConnectorTier` (`pilot` / `standard`). See connector-tier-ux-spec §2.2.
      enum: [tier-1, tier-2, tier-3]

    PipelineStageRead:
      type: object
      required: [ordinal, key, title]
      properties:
        ordinal:
          type: integer
          minimum: 1
          maximum: 8
        key:
          type: string
          description: Stable machine key for the stage
        title:
          type: string
          description: Human-readable stage label

    AdapterInstancePlaceholder:
      type: object
      required: [id, stageKey, displayName, catalogTier]
      properties:
        id:
          type: string
        stageKey:
          type: string
        displayName:
          type: string
        catalogTier:
          $ref: "#/components/schemas/CatalogTier"

    PipelineSummaryRead:
      type: object
      required: [contractVersion, stages, adapterInstances]
      properties:
        contractVersion:
          type: string
          description: OpenAPI `info.version` this read model is aligned with
          example: "0.1.0"
        stages:
          type: array
          minItems: 8
          maxItems: 8
          items:
            $ref: "#/components/schemas/PipelineStageRead"
        adapterInstances:
          type: array
          items:
            $ref: "#/components/schemas/AdapterInstancePlaceholder"

    CanalSegmentRead:
      type: object
      required: [id, kind, label, followsStageOrdinal, providerProfile]
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [buffer]
        label:
          type: string
        followsStageOrdinal:
          type: integer
          minimum: 1
          maximum: 8
        providerProfile:
          type: string
          enum: [p1-local]

    CanalSegmentsRead:
      type: object
      required: [segments]
      properties:
        segments:
          type: array
          items:
            $ref: "#/components/schemas/CanalSegmentRead"

    AdapterInstanceCreateRequest:
      type: object
      required: [catalogAdapterId]
      properties:
        catalogAdapterId:
          type: string
          description: Must match `AdapterInstancePlaceholder.id` from `GET /v1/control/pipeline`.
        operatorLabel:
          type: string
          nullable: true
          description: Optional operator-facing label (does not change catalog metadata).

    AdapterInstancePatchRequest:
      type: object
      properties:
        catalogAdapterId:
          type: string
          description: Rebind to another catalog placeholder; tier updates from catalog automatically.
        operatorLabel:
          type: string
          nullable: true

    AdapterInstanceRead:
      type: object
      required:
        [id, catalogAdapterId, catalogTier, stageKey, displayName, operatorLabel, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        catalogAdapterId:
          type: string
        catalogTier:
          $ref: "#/components/schemas/CatalogTier"
        stageKey:
          type: string
        displayName:
          type: string
        operatorLabel:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    AdapterInstanceListRead:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdapterInstanceRead"